% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_builder.R
\name{build_hypernode_yaml}
\alias{build_hypernode_yaml}
\title{Build the full configuration YAML}
\usage{
build_hypernode_yaml(eff_cfgs, global_cfg, all_bounds)
}
\arguments{
\item{eff_cfgs}{List risultante di effective_cfgs()  (override + fallback)}

\item{global_cfg}{Lista delle impostazioni globali   (serve solo per il controllo capacità)}

\item{all_bounds}{Character vector dei boundary metaboliti selezionati}
}
\value{
Stringa YAML completa
}
\description{
Build the full configuration YAML
}
