% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the epimodFBAfunctions GUI}
\usage{
run_app(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Whether to open in external browser}
}
\description{
Launch the epimodFBAfunctions GUI
}
