% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulationServer}
\alias{simulationServer}
\title{Simulation Server Module}
\usage{
simulationServer(id)
}
\arguments{
\item{id}{Module namespace id}
}
\value{
Server logic for hypernode selection and simulation config
}
\description{
Simulation Server Module
}
