% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulationUI}
\alias{simulationUI}
\title{Simulation UI Module}
\usage{
simulationUI(id)
}
\arguments{
\item{id}{Module namespace id}
}
\value{
A tabPanel for hypernode selection and simulation configuration
}
\description{
Simulation UI Module
}
