% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_builder.R
\name{write_hypernode_yaml}
\alias{write_hypernode_yaml}
\title{Write a Hypernode YAML to disk}
\usage{
write_hypernode_yaml(yaml_txt, out_dir, hypernode)
}
\arguments{
\item{yaml_txt}{A character string containing the YAML.}

\item{out_dir}{Path to the working directory.}

\item{hypernode}{Name of the hypernode (used as filename prefix).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      The full path to the file that was written.
}\if{html}{\out{</div>}}
}
\description{
Write a Hypernode YAML to disk
}
