% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bondary_conditions_builder.R
\name{writeBoundaryConditionsStatic}
\alias{writeBoundaryConditionsStatic}
\title{Write out a fully hard-coded boundary_conditions.json (for testing)}
\usage{
writeBoundaryConditionsStatic(output_json = "boundary_conditions.json")
}
\arguments{
\item{output_json}{Path to write the JSON file (default "boundary_conditions.json")}
}
\description{
Write out a fully hard-coded boundary_conditions.json (for testing)
}
